requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-decouple==3.8
openai==1.54.4
openai-whisper==20240930
moviepy==1.0.3
pydantic==2.10.2
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==24.1.0
azure-cognitiveservices-speech==1.40.0